선택 정렬 추가

+++ MySTL/Define.h
@@ -1,8 +1,8 @@
 #pragma once
 
-static const int SUCC = 0;
-static const int FAIL = 1;
+static const size_t SUCC = 0;
+static const size_t FAIL = 1;
 static const int ERROR = -1;
 
-static const int TRUE = 1;
-static const int FALSE = 0;
+static const size_t TRUE = 1;
+static const size_t FALSE = 0;
